Implemented FAN p/n detection based on RPM value
When yadro-hw-manager service is started, we check availability of the
PWM properties for the fans. Then we set the PWM to 100% and after
5 secs we read the RPM value. Based on the obtained maximum RPM value
we detect the installed fan part number.

Note: often we can't read PWM/RPM properties via dbus, therefore we
      try to read repeatedly several times.

End-user-impact: Standard and High-performance system fans are now
                 properly detected and shown in the inventory
Signed-off-by: Evgeniy Kislenok <e.kislenok@yadro.com>