storage: add tool for reflashing all MCUs
This brings `yadro-mcu-reflash` that provides a way for reflashing MCU
on all installed backplanes.

Typical usage is a call via `ExecStartPre` in `yadro-storage.service`.
But as it requires the MCU firmware images and a special config file
this is not specified as a default action and will be added by a
separate recipe on yocto side.

Signed-off-by: Alexander Filippov <a.filippov@yadro.com>